+++ packages/geoinfo/package.js
@@ -40,6 +40,7 @@ Package.onUse(function(api) {
     'client/stylesheets/popups.css',
     'client/views/markers.html',
     'client/views/panels.html',
+    'client/views/panels.js',
     'client/views/popups.html',
     'client/views/popups.js'
   ],'client');